Fix missing class loading in development
Yaml.load tries to load an object of SolidusPaypalBraintree::Response
class, but in development, if the class has never been used, it is not
in memory and Yaml.load fails.

Probably this errors does not happen in production. Anyway, I force the
require of the class here in the gem.

I hope this helps :(